Refactor: App 내의 UI 로직 분리
