+++ src/App.tsx
@@ -1,31 +1,12 @@
 import React, { useState } from "react";
-import { generateItems } from "./utils";
 import { useCallback, useMemo } from "./hooks";
 import { User, Notification } from "./types";
-import { ThemeContext, UserContext, NotificationContext } from "./context";
-import {
-  Header,
-  ItemList,
-  ComplexForm,
-  NotificationSystem,
-} from "./components";
+import { UserContext, NotificationContext, ThemeProvider } from "./context";
+import HomePage from "./pages/HomePage";
 
 const App: React.FC = () => {
-  const [theme, setTheme] = useState("light");
-  const [items, setItems] = useState(() => generateItems(1000));
   const [user, setUser] = useState<User | null>(null);
   const [notifications, setNotifications] = useState<Notification[]>([]);
-
-  const toggleTheme = useCallback(() => {
-    setTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
-  }, []);
-
-  const addItems = useCallback(() => {
-    setItems((prevItems) => [
-      ...prevItems,
-      ...generateItems(1000, prevItems.length),
-    ]);
-  }, []);
 
   const addNotification = useCallback(
     (message: string, type: Notification["type"]) => {
@@ -58,8 +39,6 @@ const App: React.FC = () => {
     addNotification("로그아웃되었습니다", "info");
   }, [addNotification]);
 
-  const themeContextValue = useMemo(() => ({ theme, toggleTheme }), [theme]);
-
   const userContextValue = useMemo(
     () => ({ user, login, logout }),
     [user, login, logout]
@@ -71,28 +50,13 @@ const App: React.FC = () => {
   );
 
   return (
-    <ThemeContext.Provider value={themeContextValue}>
+    <ThemeProvider>
       <UserContext.Provider value={userContextValue}>
         <NotificationContext.Provider value={notificationContextValue}>
-          <div
-            className={`min-h-screen ${theme === "light" ? "bg-gray-100" : "bg-gray-900 text-white"}`}
-          >
-            <Header />
-            <div className="container mx-auto px-4 py-8">
-              <div className="flex flex-col md:flex-row">
-                <div className="w-full md:w-1/2 md:pr-4">
-                  <ItemList items={items} onAddItemsClick={addItems} />
-                </div>
-                <div className="w-full md:w-1/2 md:pl-4">
-                  <ComplexForm />
-                </div>
-              </div>
-            </div>
-            <NotificationSystem />
-          </div>
+          <HomePage />
         </NotificationContext.Provider>
       </UserContext.Provider>
-    </ThemeContext.Provider>
+    </ThemeProvider>
   );
 };
