+++ src/pages/HomePage.tsx
@@ -1,0 +1,38 @@
+import { useState, useCallback } from "react";
+import { useThemeState } from "../context";
+import { ItemList, ComplexForm, NotificationSystem } from "../components";
+import { generateItems } from "../utils";
+import { Header } from "../components/Header";
+
+const HomePage: React.FC = () => {
+  const theme = useThemeState();
+
+  const [items, setItems] = useState(() => generateItems(1000));
+
+  const addItems = useCallback(() => {
+    setItems((prevItems) => [
+      ...prevItems,
+      ...generateItems(1000, prevItems.length),
+    ]);
+  }, []);
+  return (
+    <div
+      className={`min-h-screen ${theme === "light" ? "bg-gray-100" : "bg-gray-900 text-white"}`}
+    >
+      <Header />
+      <div className="container mx-auto px-4 py-8">
+        <div className="flex flex-col md:flex-row">
+          <div className="w-full md:w-1/2 md:pr-4">
+            <ItemList items={items} onAddItemsClick={addItems} />
+          </div>
+          <div className="w-full md:w-1/2 md:pl-4">
+            <ComplexForm />
+          </div>
+        </div>
+      </div>
+      <NotificationSystem />
+    </div>
+  );
+};
+
+export default HomePage;
